feat: complete Phase 0 + align with standard OSS practices
- Implement full Phase 0 local foundation (CDC, BLAKE3, encryption, vaults, snapshots, restore)
- Add comprehensive E2E tests in tests/e2e.rs (10 tests covering fidelity, dedup, encryption, history)
- Add GitHub Actions CI (fmt, clippy, test matrix on Linux/macOS/Windows + artifact uploads)
- Adopt standard OSS files and practices:
  - Dual MIT/Apache-2.0 license
  - CONTRIBUTING.md, CODE_OF_CONDUCT.md, SECURITY.md
  - rust-toolchain.toml
  - Issue & PR templates
  - .gitattributes
- Update Cargo.toml metadata, README, spec.md
- Clean up clippy warnings
- Fill placeholders in docs and use consistent 2026 dates

All tests passing (8 unit + 10 E2E). CI green.

+++ src/chunking.rs
@@ -74,8 +74,8 @@ pub fn chunk_bytes(data: &[u8], config: &ChunkConfig) -> Vec<Chunk> {
 
     let mut chunks = Vec::new();
     for entry in chunker {
-        let start = entry.offset as usize;
-        let end = start + entry.length as usize;
+        let start = entry.offset;
+        let end = start + entry.length;
         if end <= data.len() {
             let chunk_data = data[start..end].to_vec();
             chunks.push(Chunk::from_data(chunk_data));
